chore: refactor URL shortening feature and integrate nextjs-toploader
- Introduced `UrlShortenForm` component for shortening URLs with user feedback.
- Updated `package.json` and `package-lock.json` to include `nextjs-toploader` for improved loading experience.
- Refactored URL handling logic and moved URL service to `url-shorten` feature.
- Cleaned up `page.tsx` by removing old URL shortening logic and replacing it with the new form component.
- Added type definitions for URL handling in `types.ts`.
- Implemented cleanup function for expired URLs in the backend.
