+++ package.json
@@ -46,6 +46,7 @@
     "genkit": "^1.0.4",
     "lucide-react": "^0.507.0",
     "next": "15.3.1",
+    "nextjs-toploader": "^3.8.16",
     "react": "^19.0.0",
     "react-day-picker": "^9.0.0",
     "react-dom": "^19.0.0",
